Fix for http://stackoverflow.com/questions/24312622/breezejs-truncating-the-values-of-umapped-properties

+++ src/a37_dataType.js
@@ -369,7 +369,7 @@ var DataType = (function () {
             case "boolean":
                 return DataType.Boolean;
             case "number":
-                return DataType.Int32;
+                return DataType.Double;
         }
         return DataType.Undefined;
     };
